chore: Rename ScannerSection to AttendanceSection
